backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
python-dotenv==1.0.0
httpx==0.26.0
openai==1.12.0
anthropic==0.18.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pinecone-client==3.0.2
python-multipart==0.0.9
pillow==10.2.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
pydantic[email]==2.6.0

slowapi==0.1.9
resend==2.10.0
redis>=5.0.0
boto3>=1.34.0
